[SETTINGS] Wait for a camera to be detected to present the settings
The settings affect only the camera output, not the fully generared
one, so the user should have a camera connected in order to see their
effect. When the settings view is show, camera presence is detected
and if it's not available, a progressbar is shown as well as a message
telling the user to connect and power on a camera.